pyproject.toml
[tool.poetry]
name = "linkedin-scraper"
version = "0.1.0"
description = "Scrape contact information my linkedin contacts"
authors = ["stephen cyron <stephen.cyron@fdmgroup.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openpyxl = "^3.0.9"
requests-html = "^0.10.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
